Convert to common and platform modules

+++ ShaderCLikeJVM/src/main/kotlin/org/tobi29/scapes/engine/utils/shader/frontend/clike/LiteralParserJVM.kt
@@ -19,17 +19,16 @@ package org.tobi29.scapes.engine.utils.shader.frontend.clike
 import org.tobi29.scapes.engine.utils.shader.DecimalExpression
 import org.tobi29.scapes.engine.utils.shader.IntegerExpression
 import org.tobi29.scapes.engine.utils.shader.ScapesShaderParser
-import java.math.BigDecimal
-import java.math.BigInteger
+import org.tobi29.scapes.engine.utils.shader.attach
 
 internal fun ScapesShaderParser.LiteralContext.ast() = run {
     val integer = IntegerLiteral()
     if (integer != null) {
-        return@run IntegerExpression(BigInteger(integer.text))
+        return@run IntegerExpression(integer.text.toInt())
     }
     val floating = FloatingLiteral()
     if (floating != null) {
-        return@run DecimalExpression(BigDecimal(floating.text))
+        return@run DecimalExpression(floating.text.toDouble())
     }
     throw IllegalStateException("Invalid parse tree node")
 }.also { it.attach(this) }
